pyproject.toml
[project]
name = "mcp-lg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.2",
    "google-adk>=1.13.0",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.31",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
]
